requirements.txt
# DirectAdmin MCP Server Requirements

# FastAPI and Server
fastapi>=0.103.1
uvicorn>=0.23.2
starlette>=0.27.0
pydantic>=2.4.2
python-dotenv>=1.0.0

# HTTP Client and SSE
httpx>=0.25.0
aiohttp>=3.8.5
sseclient-py>=1.7.2

# MCP SDK
mcp>=0.1.0

# Utilities
python-multipart>=0.0.6
typing-extensions>=4.8.0